requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
flask-login==0.6.3
flask-mail==0.9.1
google-generativeai==0.3.2
line-bot-sdk==3.7.0
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.21
werkzeug==2.3.7
psycopg2-binary==2.9.7
